# backend/openapi/openapi.json
{
  "components": {
    "schemas": {
      "AuthResponse": {
        "properties": {
          "user": {
            "$ref": "#/components/schemas/User"
          }
        },
        "required": [
          "user"
        ],
        "type": "object"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              }
            },
            "required": [
              "code",
              "message"
            ],
            "type": "object"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
      "Form": {
        "properties": {
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "description": {
            "nullable": true,
            "type": "string"
          },
          "fields": {
            "items": {
              "$ref": "#/components/schemas/FormField"
            },
            "type": "array"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "publicSlug": {
            "nullable": true,
            "type": "string"
          },
          "publicUrl": {
            "nullable": true,
            "type": "string"
          },
          "publishedAt": {
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "status": {
            "enum": [
              "draft",
              "published"
            ],
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "status",
          "fields",
          "createdAt",
          "updatedAt"
        ],
        "type": "object"
      },
      "FormField": {
        "properties": {
          "config": {
            "additionalProperties": true,
            "type": "object"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "label": {
            "type": "string"
          },
          "options": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "placeholder": {
            "nullable": true,
            "type": "string"
          },
          "position": {
            "type": "integer"
          },
          "required": {
            "type": "boolean"
          },
          "type": {
            "enum": [
              "text",
              "textarea",
              "email",
              "number",
              "select",
              "checkbox"
            ],
            "type": "string"
          }
        },
        "required": [
          "id",
          "position",
          "type",
          "label",
          "required",
          "options",
          "config"
        ],
        "type": "object"
      },
      "FormFieldInput": {
        "properties": {
          "config": {
            "additionalProperties": true,
            "type": "object"
          },
          "label": {
            "maxLength": 160,
            "minLength": 1,
            "type": "string"
          },
          "options": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "placeholder": {
            "nullable": true,
            "type": "string"
          },
          "required": {
            "type": "boolean"
          },
          "type": {
            "enum": [
              "text",
              "textarea",
              "email",
              "number",
              "select",
              "checkbox"
            ],
            "type": "string"
          }
        },
        "required": [
          "type",
          "label"
        ],
        "type": "object"
      },
      "FormListResponse": {
        "properties": {
          "forms": {
            "items": {
              "$ref": "#/components/schemas/Form"
            },
            "type": "array"
          }
        },
        "required": [
          "forms"
        ],
        "type": "object"
      },
      "FormRequest": {
        "properties": {
          "description": {
            "nullable": true,
            "type": "string"
          },
          "fields": {
            "items": {
              "$ref": "#/components/schemas/FormFieldInput"
            },
            "type": "array"
          },
          "title": {
            "maxLength": 160,
            "minLength": 1,
            "type": "string"
          }
        },
        "required": [
          "title"
        ],
        "type": "object"
      },
      "HealthResponse": {
        "properties": {
          "status": {
            "example": "ok",
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "type": "object"
      },
      "LoginRequest": {
        "properties": {
          "email": {
            "format": "email",
            "type": "string"
          },
          "password": {
            "format": "password",
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "type": "object"
      },
      "RegisterRequest": {
        "properties": {
          "email": {
            "format": "email",
            "type": "string"
          },
          "name": {
            "minLength": 1,
            "type": "string"
          },
          "password": {
            "format": "password",
            "maxLength": 72,
            "minLength": 8,
            "type": "string"
          }
        },
        "required": [
          "name",
          "email",
          "password"
        ],
        "type": "object"
      },
      "User": {
        "properties": {
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "email": {
            "format": "email",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "email",
          "name",
          "createdAt"
        ],
        "type": "object"
      }
    },
    "securitySchemes": {
      "sessionCookie": {
        "in": "cookie",
        "name": "form_builder_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "Form Builder API",
    "version": "0.1.0"
  },
  "openapi": "3.0.3",
  "paths": {
    "/api/auth/login": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            },
            "description": "Authenticated user"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid credentials"
          }
        },
        "summary": "Login with email and password"
      }
    },
    "/api/auth/logout": {
      "post": {
        "responses": {
          "204": {
            "description": "Session revoked"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Logout current session"
      }
    },
    "/api/auth/me": {
      "get": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            },
            "description": "Authenticated user"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthenticated"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Get current authenticated user"
      }
    },
    "/api/auth/register": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            },
            "description": "Authenticated user"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid request"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Email already registered"
          }
        },
        "summary": "Register with email and password"
      }
    },
    "/api/forms": {
      "get": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FormListResponse"
                }
              }
            },
            "description": "Forms list"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthenticated"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "List authenticated user's forms"
      },
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FormRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Form"
                }
              }
            },
            "description": "Created form"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthenticated"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Create a form draft"
      }
    },
    "/api/forms/{formId}": {
      "delete": {
        "parameters": [
          {
            "description": "Form ID",
            "in": "path",
            "name": "formId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Form deleted"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthenticated"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Form not found"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Delete a form owned by the authenticated user"
      },
      "get": {
        "parameters": [
          {
            "description": "Form ID",
            "in": "path",
            "name": "formId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Form"
                }
              }
            },
            "description": "Form"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthenticated"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Form not found"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Get a form owned by the authenticated user"
      },
      "put": {
        "parameters": [
          {
            "description": "Form ID",
            "in": "path",
            "name": "formId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FormRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Form"
                }
              }
            },
            "description": "Updated form"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthenticated"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Form not found"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Update a form owned by the authenticated user"
      }
    },
    "/api/forms/{formId}/publish": {
      "post": {
        "parameters": [
          {
            "description": "Form ID",
            "in": "path",
            "name": "formId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Form"
                }
              }
            },
            "description": "Published form"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthenticated"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Form not found"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Publish a form and expose a public slug"
      }
    },
    "/api/forms/{formId}/unpublish": {
      "post": {
        "parameters": [
          {
            "description": "Form ID",
            "in": "path",
            "name": "formId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Form"
                }
              }
            },
            "description": "Draft form"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthenticated"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Form not found"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Return a published form to draft status"
      }
    },
    "/api/public/forms/{slug}": {
      "get": {
        "parameters": [
          {
            "description": "Public form slug",
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Form"
                }
              }
            },
            "description": "Published form"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Form not found"
          }
        },
        "summary": "Get a published form by public slug"
      }
    },
    "/healthz": {
      "get": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            },
            "description": "Service is healthy"
          }
        },
        "summary": "Health check"
      }
    },
    "/readyz": {
      "get": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            },
            "description": "Database connection is ready"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            },
            "description": "Database connection is unavailable"
          }
        },
        "summary": "Readiness check"
      }
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ]
}